Cache Rates and Refresh them Periodically
AlphaVantages API is, as many others, rate-limited. This introduces a
local cache of data in all of the spec's desired currencies, in order
to guarantee that the User will at least see something on her screen.

To simplify things, we use SuckerPunch and a Concurrent::TaskTimer to
perform the refresh as a scheduled background job, even if (by using
Tasks in the :io pool) we're already doing things in other threads.